test: share color test helpers
